上下架读取商品名称

+++ src/main/java/org/jeecgframework/core/common/dao/jdbc/SimpleJdbcTemplate.java
@@ -96,7 +96,7 @@ public class SimpleJdbcTemplate {
 			}
 
 		}catch (Exception e) {
-			return 0;
+			return (long)0;
 		}
 	}
